+++ TealiumSampleApp/App.js
@@ -19,7 +19,6 @@ let allTests = [
           "anotherkey": "anotherval",
           "test_int": 1
         });
-        console.log("HELLO WORLD")
         Tealium.trackEventForInstanceName("instance-2", "test_event_2");
         Tealium.getUserConsentStatusForInstanceName("instance-2", function(userConsentStatus) {
            console.log("consent status 'instance-2': " + userConsentStatus);
@@ -33,13 +32,11 @@ let allTests = [
     title: "Track View",
     run: () => {
       try {
-        console.log("HEEELLLLOOO")
         Tealium.trackView("test_view", {
           "title": "test_view",
           "event_title": "test_view",
           "testkey": "testval",
-          "anotherkey": "anotherval",
-          "test_int": 1
+          "anotherkey": "anotherval"
         });
         Tealium.trackViewForInstanceName("instance-2", "test_view_2");
       } catch(err) {
@@ -358,7 +355,7 @@ export default class App extends React.Component {
     Tealium.initializeCustom(
       'tealiummobile', 'react-native', 'qa',
       'your-ios-datasource', 'your-android-datasource', 'instance-2',
-      true, null, null, "https://collect.tealiumiq.com/event", true
+      true, null, null, "https://some.other.server.com/endpoint", true
     );
   }